Retire the flat IR: the scope tree is the only program representation
Deletes, now that explain_tree removed the last reader:
- explain.rs (the flat rewrite, builder, clone, reverse rules) — its tree
  port is explain_tree; the arity transfer functions move to ir.rs.
- The flat Lowering and lower() in lower.rs (lower_tree remains).
- ir::Node / ir::Program and its impl (dump, depths, optimize, rewrite,
  validate_lift_iter). ir.rs is now the shared row/op vocabulary: LinearOp,
  RowLike, field/condition evaluation, arity transfer.
- render_program and the FLAT=1 branches in both backends; the tree path
  (regions, optimizer, --explain via explain_tree) is the only path.

Noted: the flat validate_lift_iter discipline check (LiftIter result must
not be referenced in its own scope) has no tree equivalent yet; LiftIter is
rewrite-emitted today, with only the applicative parser able to surface it.

Verified after the sweep: 15 lib tests, corpus output counts unchanged on
both backends, --explain demand sets unchanged (scc pairs query).
