--- interactive/src/explain.rs
@@ -1,1423 +1,0 @@
-//! DDIR-to-DDIR explanation rewrite.
-//!
-//! `explain(p)` transforms a Program into one whose execution produces
-//! per-input demand-set explanations for queries against the original program's result.
-//!
-//! ### Architecture (per design notes)
-//!
-//! ```text
-//! -- INPUT:
-//! --    a. a user dataflow,
-//! --    b. live input sources,
-//! --    c. live output queries.
-//! --
-//! -- This module is a pure IR-to-IR transform: it returns a Program. Any DDIR
-//! -- backend (`ddir_vec`, `ddir_col`) then executes that Program; running it on
-//! -- the inputs + a stream of output queries produces, for each output query,
-//! -- a subset of the data sources that "explains" the query — running the user
-//! -- dataflow on that subset reproduces the queried output. The explanation
-//! -- updates live as the data sources and queries change.
-//! --
-//! -- The returned Program contains two clones of the user dataflow plus a
-//! -- reverse-tracing dataflow:
-//! --
-//! -- The *witness* clone runs on the actual data sources and produces the
-//! -- reference output the explanation must reproduce.
-//! --
-//! -- Alongside, in an iterative scope `explain`, the *forward* clone runs the
-//! -- user dataflow on the data sources restricted to per-input *demand-sets*.
-//! -- Demand-sets start empty and grow until the forward clone reproduces every
-//! -- queried output.
-//! --
-//! -- The demand-sets are populated by a reverse-tracing dataflow. Each output
-//! -- query enters the reverse dataflow as an "input" to it; it traverses each
-//! -- operator in turn, asking that operator's inputs for the updates that
-//! -- could play a role in forming the queried output.
-//! --
-//! --        IN0 \                       d_IN0 <-\
-//! --             OP -> OUT    becomes            OP^d <- d_OUT
-//! --        IN1 /                       d_IN1 <-/
-//! --
-//! -- The unit of demand is `(data, time)`, where `time` is the user dataflow's
-//! -- iteration coordinate(s) nested under the ambient (host) time. Time is
-//! -- load-bearing: an input may only "explain" an output if it came before it.
-//! --
-//! -- Both clones use `clone_with_lifts`, which inserts `lift_iter` at every
-//! -- scope exit so that every collection inside a user scope has a "host-
-//! -- visible" form at the outer (explain) scope with its user-iter coords
-//! -- folded into the value. The `CloneResult.host[id]` map exposes this
-//! -- outer-scope id for every node; the reverse rules read from it to build
-//! -- their witness pair tables.
-//! --
-//! -- The result is one demand-set per input — the subset of input rows that,
-//! -- taken together, reproduce every queried output.
-//!
-//! witness: { <user dataflow, cloned with lift_iter at scope exits> }
-//!
-//! explain: {
-//!     -- per-input demand-set variables; grown by `reverse` below
-//!     -- to whatever proves sufficient to reproduce the demanded outputs.
-//!
-//!     forward: { <user dataflow, inputs restricted to demand-sets, lifted> }
-//!     reverse: { backwards trace from output queries to demand-sets }
-//! }
-//! ```
-//!
-//! ### Vocabulary
-//!
-//! Three related but distinct concepts share the "demand" word:
-//!
-//! - **demand-set** (`demand_set_<i>`, one per Input): the accumulating
-//!   per-input subset — the *result* of explanation.
-//! - **demand variable** (`demand_<N>`, one per non-trivial IR node): an
-//!   intermediate feedback variable through which the reverse-tracing
-//!   dataflow propagates demand from outputs back to inputs.
-//! - **demand row**: a single record `(data, time, q)` flowing through
-//!   the reverse dataflow — "output query `q` needs `data` produced at
-//!   `time` from this point."
-//!
-//! For an IR node `N` at user-scope depth `D`, a demand row in `demand_<N>`
-//! has shape:
-//!
-//! ```text
-//! (K_N ; V_N ++ user_chain[D] ++ [q])
-//! ```
-//!
-//! where `user_chain[D]` is one i64 per enclosing user scope (innermost
-//! first; the position lifted by the innermost scope's `lift_iter` comes
-//! first). `q` is the query id.
-//!
-//! ### Out of scope for this pass
-//!
-//! - `Reduce::Count` — handled by the same keyed lookup as Min/Distinct,
-//!   but without value-narrowing. A real Count-flavored rule would need to
-//!   reason about which input rows summed to the demanded count.
-//! - Multi-op `Linear` chains. `emit_reverse` currently panics on these;
-//!   the rewrite assumes the optimizer leaves Linear nodes single-op.
-//! - Pre-optimization. The optimizer can clean up dead lifts and unused
-//!   nodes after the rewrite emits them.
-//!
-//! ### Future investigation
-//!
-//! **Self-protection at non-pass-through ops.** Today, Concat is the one
-//! pass-through-eligible op that *isn't* pass-through, because its per-input
-//! SP lookup against `host[input]` is what filters spurious demand routed
-//! to consumers downstream of the Concat — specifically, demand at user_chain
-//! values that the consumer never actually emitted. The Join's right contrib
-//! is one example: it filters on `u_R ≤ u_out` (its own pair's right side),
-//! but with static edges where `u_R = 0`, that filter admits demand at any
-//! `u_out ≥ 0` — including u_out values the Join never produced output at.
-//! The Concat's per-input SP catches this externally.
-//!
-//! A cleaner model would shift the responsibility inward: every
-//! non-pass-through op (Reduce, Join, Linear[Project]) self-filters its
-//! incoming `dep_y` against `host[<own-output>]` at the start of its bward,
-//! asking "is this dep row consistent with what I actually emitted?"
-//! Concat would then become pass-through. Each op becomes responsible for
-//! receiving only valid demand, no externality on its consumers.
-//!
-//! Trade-off: a self-filter adds ~5 IR nodes per non-pass-through op vs.
-//! the current ~5 nodes per Concat input. Programs with many ops behind one
-//! Concat win; programs with few Concats and many ops lose. The diagnostic
-//! value is the bigger gain — fewer "action at a distance" relationships
-//! to reason about.
-//!
-//! ## Extending the IR — touchpoints
-//!
-//! When adding a new variant to one of the three IR enums, the type system
-//! will catch most omissions because the semantic dispatch sites (those
-//! enumerated below as "compile-error sites") are exhaustive matches. The
-//! list still helps a contributor know *why* each site needs an update.
-//!
-//! ### New `Node` variant
-//!
-//! - `ir.rs`: define the variant; update `Program::depths()` if the node
-//!   enters/leaves a scope; update the IR printer (`Display`/`dump`).
-//! - `lower.rs`, `parse/`: lowering + parser support, if user-facing.
-//! - `explain.rs::arities::compute_arities`: produce the new node's
-//!   `(k, v)` shape. *Compile-error site.*
-//! - `explain.rs::clone::clone_with_lifts`: emit the variant in the
-//!   witness/forward clones; decide if `host[N]` aliases an input (like
-//!   `Arrange`/`Leave`) or stands alone. *Two compile-error sites: the
-//!   control-flow vs data-node dispatch above, and the emission match.*
-//! - `explain.rs::reverse::emit_reverse`: *the backward rule.* Decide:
-//!   - pass-through (data and scope unchanged — `Arrange`, `Inspect`)?
-//!   - host-lookup against the input's pair table (`Reduce`-style,
-//!     `Linear[Project]`-style, `Concat`-per-input)?
-//!   - special-purpose (`Leave` injects user_chain; `Bind` advances iter)?
-//!   - Does it cross a depth boundary? Only `Leave` does today; if your
-//!     node does too, mirror its structure (and look at the host vs dep
-//!     user-chain split on `Side`).
-//!   *Compile-error site.*
-//!
-//! ### New `LinearOp` variant
-//!
-//! - `ir.rs`: define; printer.
-//! - `parse/`, `lower.rs`: parser + lowering if user-facing.
-//! - `explain.rs::arities::apply_ops_arity`: how the op changes `(k, v)`.
-//!   *Compile-error site.*
-//! - `explain.rs::reverse::emit_reverse`'s `Linear` arm: backward rule.
-//!   *Compile-error site.*
-//!
-//! ### New `FieldExpr` variant
-//!
-//! - `ir.rs`: define; `eval_field_into`, `eval_field_raw` (forward eval).
-//!   *Compile-error sites.*
-//! - `parse/`: syntax, optional. (`Sub` is rewrite-emitted only today — no
-//!   parser entry; users can't write `$1[0] - 1`.)
-//! - `explain.rs::analyze_lossy_invertibility`: classify as recoverable
-//!   (drives the pure-map shortcut) or opaque (falls back to the
-//!   pair-table path — sound but lossier). *Compile-error site.*
-//! - `explain.rs::expand_pos_one` (used inside `emit_lookup_join`):
-//!   propagate the variant through `Pos` expansion. *Compile-error site.*
-
-use std::collections::BTreeMap;
-
-use crate::ir::{Id, Node, Program};
-use crate::parse::{FieldExpr, Projection, Reducer};
-
-use builder::Builder;
-use arities::compute_arities;
-use clone::CloneResult;
-
-/// Transform a `Program` into one whose execution produces per-input
-/// demand-set explanations for queries against the original's result.
-/// See the module doc for the architecture.
-///
-/// `input_arities` gives `(key_arity, val_arity)` per positional input;
-/// `import_arities` gives the same per named import (entries needed for
-/// every distinct `Import { name }` referenced in `p`). Both are
-/// necessary because data-source shapes aren't recoverable from the IR
-/// alone (Projections only invert with known input arity).
-pub fn explain(
-    p: &Program,
-    input_arities: &[(usize, usize)],
-    import_arities: &BTreeMap<String, (usize, usize)>,
-) -> Program {
-    let mut b = Builder::new();
-    let arities = compute_arities(p, input_arities, import_arities);
-    let depths = p.depths();
-    // The two user-chain lengths we track at each node:
-    //
-    // * `host_user_lens[N]` — how many user-iter coords the lifted, host-
-    //   visible collection at `N` carries in val. For most nodes this is
-    //   `depths[N]`; for `Leave(inner, _)`, host[Leave] aliases host[inner],
-    //   so the host length equals `depths[inner]` (one greater than
-    //   `Leave`'s own depth). Used only for *lookups* against `host[N]` —
-    //   i.e. via `Side`'s pair construction.
-    //
-    // * `dep_user_lens[N]` — how many user-iter coords the *demand* at `N`
-    //   carries in val. This is always `depths[N]` — no override.
-    //
-    // The split matters at `Leave`: dep[Leave] lives in the outer scope,
-    // so it carries the outer scope's user_chain length, even though
-    // host[Leave] carries the (longer) inner one. `Leave`'s bward is the
-    // place that *injects* the extra inner user_chain coord (via an SP
-    // lookup against host[inner]) — no other node has to know about the
-    // depth boundary.
-    let host_user_lens: BTreeMap<Id, usize> = p.nodes.iter().map(|(&id, node)| {
-        let len = if let Node::Leave(inner, _) = node {
-            *depths.get(inner).unwrap_or(&0)
-        } else {
-            *depths.get(&id).unwrap_or(&0)
-        };
-        (id, len)
-    }).collect();
-    let dep_user_lens: &BTreeMap<Id, usize> = &depths;
-    let n_inputs = input_arities.len();
-
-    // Distinct import names referenced by `p`, in deterministic order.
-    let import_names: Vec<String> = {
-        let mut s = std::collections::BTreeSet::new();
-        for node in p.nodes.values() {
-            if let Node::Import { name } = node { s.insert(name.clone()); }
-        }
-        s.into_iter().collect()
-    };
-
-    // ---- outer scope ----
-    // Original inputs of `p`, one outer-scope Import per referenced name,
-    // plus one extra "query" input.
-    let original_inputs: Vec<Id> = (0..n_inputs).map(|i| b.input(i)).collect();
-    let original_imports: BTreeMap<String, Id> = import_names.iter()
-        .map(|n| (n.clone(), b.push(Node::Import { name: n.clone() })))
-        .collect();
-    let query_input = b.input(n_inputs);
-
-    // witness: a clone of `p`, with lift_iter chains so every witness
-    // collection has a host-visible `(data, user)` form via auto-leave at
-    // each enclosing user scope's exit.
-    let witness = b.clone_with_lifts(p, &original_inputs, &original_imports, 0);
-
-    // ---- explain scope ----
-    b.scope_open();
-
-    // Demand-set Variables (one per input, one per import).
-    let demand_sets: Vec<Id> = (0..n_inputs).map(|_| b.variable()).collect();
-    let import_demand_sets: BTreeMap<String, Id> = import_names.iter()
-        .map(|n| (n.clone(), b.variable()))
-        .collect();
-
-    // forward inputs: demand_set_<i> | semijoin(actual_input_<i>).
-    // Enter actual inputs into explain scope implicitly; semijoin restricts to
-    // demanded rows.
-    let forward_inputs: Vec<Id> = (0..n_inputs)
-        .map(|i| {
-            let (k, v) = input_arities[i];
-            b.semijoin_data(demand_sets[i], original_inputs[i], k, v)
-        })
-        .collect();
-    let forward_imports: BTreeMap<String, Id> = import_names.iter()
-        .map(|n| {
-            let (k, v) = import_arities[n];
-            let semi = b.semijoin_data(import_demand_sets[n], original_imports[n], k, v);
-            (n.clone(), semi)
-        })
-        .collect();
-
-    // forward: same clone procedure as witness, with substituted inputs.
-    // Offset = 1 because this clone lives INSIDE the explain scope: its real
-    // PointStamp depth at any point is one more than its local user_level.
-    let forward = b.clone_with_lifts(p, &forward_inputs, &forward_imports, 1);
-
-    // Demand Variables are pre-allocated *only* for user-program `var` IR
-    // nodes (`Node::Variable`). These are the only places where the demand
-    // chain has a structurally-required feedback (forward cycles around
-    // user `var`s induce backward cycles that need a Variable to close).
-    //
-    // For every other node, demand is a closed-form expression computed
-    // in the reverse walk below.
-    let mut demand_var: BTreeMap<Id, Id> = BTreeMap::new();
-    for (&id, node) in &p.nodes {
-        if matches!(node, Node::Variable) {
-            demand_var.insert(id, b.variable());
-        }
-    }
-
-    // reverse rules. Walk nodes in *reverse* id order so each node's
-    // `contribs[id]` is already populated by its higher-id consumers by
-    // the time we reach it. For each node we compute its demand expression
-    // from `contribs[id]`, store it in `demand_var`, then dispatch the
-    // node's bward rule (which pushes onto its inputs' contribs).
-    //
-    // Query input directly seeds `contribs[first_export]` — the demand
-    // starts with the query rows against the first export. v0 only
-    // explains a single output; multi-export programs would need one
-    // query input per export and per-output dispatch in the seeding.
-    let primary_export = p.export.first()
-        .expect("explain: program has no export to seed query input against").1;
-    let mut contribs: BTreeMap<Id, Vec<Id>> = BTreeMap::new();
-    contribs.entry(primary_export).or_default().push(query_input);
-
-    for (&id, node) in p.nodes.iter().rev() {
-        // Scope / EndScope carry no demand and have no bward action.
-        if matches!(node, Node::Scope | Node::EndScope) { continue; }
-        // Bind has no demand of its own; its handler in `emit_reverse`
-        // routes the variable's pre-allocated demand into the value's
-        // contribs.
-        if matches!(node, Node::Bind { .. }) {
-            b.emit_reverse(id, node, &witness, &forward, &demand_var, &arities, &host_user_lens, dep_user_lens, &mut contribs);
-            continue;
-        }
-        let cs = contribs.remove(&id).unwrap_or_default();
-        if cs.is_empty() {
-            // No upstream demand. For a user `var`, still bind the
-            // pre-allocated Variable to itself so the feedback edge is
-            // structurally valid (it stays empty in practice).
-            if matches!(node, Node::Variable) {
-                let var = demand_var[&id];
-                b.bind(var, var);
-            }
-            continue;
-        }
-        let combined = if cs.len() == 1 { cs[0] } else { b.concat(cs) };
-        let (k, v) = arities[&id];
-        let user_len = dep_user_lens[&id];
-        let val_arity = v + user_len + 1; // V + user_chain + [q]
-        let dist = b.distinct_full(combined, k, val_arity);
-        b.debug_inspect(dist, format!("demand_{}", id));
-        if matches!(node, Node::Variable) {
-            // User `var`: bind the pre-allocated Variable to its demand
-            // expression. `demand_var[id]` already holds the Variable id
-            // (set above); leave it as-is so consumers reference the
-            // Variable, not the body expression directly.
-            let var = demand_var[&id];
-            b.bind(var, dist);
-        } else {
-            // Non-user-var: demand is the expression itself, no Variable
-            // wrapping. Inputs that consume this node read the expression
-            // id directly.
-            demand_var.insert(id, dist);
-        }
-        b.emit_reverse(id, node, &witness, &forward, &demand_var, &arities, &host_user_lens, dep_user_lens, &mut contribs);
-    }
-
-    // Bind demand-set variables for inputs and imports symmetrically:
-    //   demand_set_X := distinct(demand_set_X + (demand_<X> | strip | semijoin actual)).
-    // Build a Vec mapping input index `i` to its IR id in `p`, so the
-    // per-input loop below is O(n) total instead of O(n^2). Imports are
-    // looked up by name; multiple `Import { name }` nodes in `p` share
-    // demand via the dedup pass — we route from any one of them.
-    let mut input_ids: Vec<Option<Id>> = vec![None; n_inputs];
-    let mut import_ids: BTreeMap<String, Id> = BTreeMap::new();
-    for (&id, node) in &p.nodes {
-        match node {
-            Node::Input(i) => { input_ids[*i] = Some(id); }
-            Node::Import { name } => { import_ids.entry(name.clone()).or_insert(id); }
-            _ => {}
-        }
-    }
-    for i in 0..n_inputs {
-        let in_id = input_ids[i].expect("input not found in program");
-        let (kx, vx) = arities[&in_id];
-        // Inputs are always at depth 0 → user_chain is empty.
-        let stripped = b.project(demand_var[&in_id], strip_user_and_q(kx, vx));
-        let semi = b.semijoin_data(stripped, original_inputs[i], kx, vx);
-        let combined = b.concat(vec![demand_sets[i], semi]);
-        let dist = b.distinct_full(combined, kx, vx);
-        b.debug_inspect(dist, format!("demand_set:{}", i));
-        b.bind(demand_sets[i], dist);
-    }
-    for name in &import_names {
-        let imp_id = import_ids[name];
-        let (kx, vx) = arities[&imp_id];
-        let stripped = b.project(demand_var[&imp_id], strip_user_and_q(kx, vx));
-        let semi = b.semijoin_data(stripped, original_imports[name], kx, vx);
-        let combined = b.concat(vec![import_demand_sets[name], semi]);
-        let dist = b.distinct_full(combined, kx, vx);
-        b.bind(import_demand_sets[name], dist);
-    }
-
-    // Per-source demand-set ids inside the explain scope; we leave each
-    // out and register as a named export after closing the scope.
-    let inputs_leaves: Vec<(String, Id)> = demand_sets.iter().enumerate()
-        .map(|(i, &mv)| (format!("demand:input{}", i), mv))
-        .collect();
-    let imports_leaves: Vec<(String, Id)> = import_names.iter()
-        .map(|n| (format!("demand:{}", n), import_demand_sets[n]))
-        .collect();
-
-    b.scope_close();
-    for (name, inner) in inputs_leaves.into_iter().chain(imports_leaves) {
-        let outer = b.leave(inner, 1);
-        b.add_export(name, outer);
-    }
-    b.into_program()
-}
-
-/// Thin builder wrapper around `Program` for incremental IR construction.
-///
-/// Each push method appends a node and returns its fresh id. Composite
-/// constructors (`reduce`, `join`) emit the implicit `Arrange` wrappers the
-/// IR requires of those ops, so callers can think in terms of collections.
-mod builder {
-    use std::collections::BTreeMap;
-
-    use crate::ir::{Id, LinearOp, Node, Program};
-    use crate::parse::{Condition, Projection, Reducer};
-
-    pub struct Builder {
-        program: Program,
-        next_id: Id,
-    }
-
-    impl Builder {
-        pub(super) fn new() -> Self {
-            Builder {
-                program: Program { nodes: BTreeMap::new(), export: Vec::new() },
-                next_id: 0,
-            }
-        }
-        pub(super) fn push(&mut self, n: Node) -> Id {
-            let id = self.next_id;
-            self.next_id += 1;
-            self.program.nodes.insert(id, n);
-            id
-        }
-        pub(super) fn input(&mut self, n: usize) -> Id { self.push(Node::Input(n)) }
-        pub(super) fn variable(&mut self) -> Id { self.push(Node::Variable) }
-        pub(super) fn arrange(&mut self, input: Id) -> Id { self.push(Node::Arrange(input)) }
-        pub(super) fn linear(&mut self, input: Id, ops: Vec<LinearOp>) -> Id {
-            self.push(Node::Linear { input, ops })
-        }
-        pub(super) fn project(&mut self, input: Id, p: Projection) -> Id {
-            self.linear(input, vec![LinearOp::Project(p)])
-        }
-        pub(super) fn filter(&mut self, input: Id, c: Condition) -> Id {
-            self.linear(input, vec![LinearOp::Filter(c)])
-        }
-        pub(super) fn concat(&mut self, ids: Vec<Id>) -> Id { self.push(Node::Concat(ids)) }
-        pub(super) fn reduce(&mut self, input: Id, r: Reducer) -> Id {
-            let arr = self.arrange(input);
-            self.push(Node::Reduce { input: arr, reducer: r })
-        }
-        pub(super) fn join(&mut self, left: Id, right: Id, p: Projection) -> Id {
-            let l = self.arrange(left);
-            let r = self.arrange(right);
-            self.push(Node::Join { left: l, right: r, projection: p })
-        }
-        pub(super) fn inspect(&mut self, input: Id, label: String) -> Id {
-            self.push(Node::Inspect { input, label })
-        }
-        pub(super) fn leave(&mut self, inner: Id, scope_level: usize) -> Id {
-            self.push(Node::Leave(inner, scope_level))
-        }
-        pub(super) fn bind(&mut self, variable: Id, value: Id) {
-            self.push(Node::Bind { variable, value });
-        }
-        pub(super) fn scope_open(&mut self) { self.push(Node::Scope); }
-        pub(super) fn scope_close(&mut self) { self.push(Node::EndScope); }
-        pub(super) fn add_export(&mut self, name: String, id: Id) { self.program.export.push((name, id)); }
-        pub(super) fn into_program(self) -> Program { self.program }
-    }
-}
-
-/// Per-IR-node (key_arity, val_arity) inference.
-///
-/// Walks the program to a fixed point, deriving each node's data-shape from
-/// its inputs and the op-specific rule. Needed because input shapes aren't
-/// recoverable from the IR alone — `Projection`s only invert with known
-/// input arity, and lift_iter sites need to know how many user-iter coords
-/// already sit in the val.
-pub mod arities {
-    use std::collections::BTreeMap;
-
-    use crate::ir::{Id, LinearOp, Node, Program};
-    use crate::parse::{FieldExpr, Reducer};
-
-    pub fn compute_arities(
-        p: &Program,
-        input_arities: &[(usize, usize)],
-        import_arities: &BTreeMap<String, (usize, usize)>,
-    ) -> BTreeMap<Id, (usize, usize)> {
-        // Variables are referenced before their Binds appear in id order;
-        // resolve a Variable's shape via its body.
-        let var_body: BTreeMap<Id, Id> = p.nodes.iter().filter_map(|(_, n)| {
-            if let Node::Bind { variable, value } = n {
-                Some((*variable, *value))
-            } else { None }
-        }).collect();
-
-        let mut out: BTreeMap<Id, (usize, usize)> = BTreeMap::new();
-        loop {
-            let before = out.len();
-            for (&id, node) in &p.nodes {
-                if out.contains_key(&id) { continue; }
-                let shape = match node {
-                    Node::Input(i) => Some(input_arities[*i]),
-                    Node::Import { name } => Some(
-                        *import_arities.get(name)
-                            .unwrap_or_else(|| panic!("explain: no arity registered for import {:?}", name))
-                    ),
-                    Node::Linear { input, ops } => out.get(input).map(|s| apply_ops_arity(*s, ops)),
-                    // Try each input — for self-recursive Variables that appear
-                    // as `Concat([var, ...])`, the first input's shape isn't
-                    // known on early passes; pick any input that has a known
-                    // shape and let fixed-point iteration propagate.
-                    Node::Concat(ids) => ids.iter().find_map(|i| out.get(i).copied()),
-                    Node::Arrange(input) => out.get(input).copied(),
-                    // A projection's arity is the sum of each field's *width*,
-                    // not the count of field-exprs: `Pos(r)` is a whole-row
-                    // reference that expands to input row `r`'s arity. The
-                    // join's input rows are [key, left_val, right_val].
-                    Node::Join { left, right, projection } => match (out.get(left), out.get(right)) {
-                        (Some(&(kl, vl)), Some(&(_kr, vr))) => {
-                            let rows = [kl, vl, vr];
-                            Some((proj_arity(&projection.key, &rows), proj_arity(&projection.val, &rows)))
-                        }
-                        _ => None,
-                    },
-                    Node::Reduce { input, reducer } => out.get(input).map(|s| match reducer {
-                        Reducer::Distinct => (s.0, 0),
-                        Reducer::Min => (s.0, s.1),
-                        Reducer::Count => (s.0, 1),
-                    }),
-                    Node::Inspect { input, .. } => out.get(input).copied(),
-                    Node::Leave(inner, _) => out.get(inner).copied(),
-                    Node::Variable => var_body.get(&id).and_then(|v| out.get(v).copied()),
-                    Node::Scope | Node::EndScope | Node::Bind { .. } => None,
-                };
-                if let Some(s) = shape { out.insert(id, s); }
-            }
-            if out.len() == before { break; }
-        }
-        out
-    }
-
-    pub(crate) fn apply_ops_arity((mut k, mut v): (usize, usize), ops: &[LinearOp]) -> (usize, usize) {
-        for op in ops {
-            match op {
-                // Project's input rows are [key, val]; expand `Pos` refs to
-                // their row arities rather than counting field-exprs.
-                LinearOp::Project(p) => {
-                    let rows = [k, v];
-                    k = proj_arity(&p.key, &rows);
-                    v = proj_arity(&p.val, &rows);
-                }
-                LinearOp::Filter(_) | LinearOp::Negate | LinearOp::EnterAt(_) => {}
-                LinearOp::LiftIter => { v += 1; }
-            }
-        }
-        (k, v)
-    }
-
-    /// Width (output columns) a single `FieldExpr` expands to, given the
-    /// arities of the input rows it may reference. `Pos(r)` is a whole-row
-    /// reference of width `rows[r]`; index/const are single columns.
-    fn field_width(f: &FieldExpr, rows: &[usize]) -> usize {
-        match f {
-            FieldExpr::Pos(r) => rows.get(*r).copied().unwrap_or(0),
-            FieldExpr::Index(_, _) | FieldExpr::Const(_) => 1,
-            FieldExpr::Neg(inner) => field_width(inner, rows),
-            FieldExpr::Sub(a, _) => field_width(a, rows),
-        }
-    }
-
-    /// Total arity of one projection side (`key`/`val`): the sum of its
-    /// fields' widths.
-    pub(crate) fn proj_arity(fields: &[FieldExpr], rows: &[usize]) -> usize {
-        fields.iter().map(|f| field_width(f, rows)).sum()
-    }
-}
-
-/// IR cloning with implicit lift_iter at scope exits.
-///
-/// `clone_with_lifts` emits `p` into `b` with a `lift_iter` chain at every
-/// scope exit, so every non-Input data node has a host-visible version
-/// reachable as `host[id]`. Inputs are aliased to `input_subst`.
-///
-/// Convention: the lifted user-iter coords are appended innermost-first. A
-/// node at depth D has D fields of user time appended to val in its host-
-/// visible form. `enclosing_scope_depth` adjusts every emitted `Leave` to
-/// account for an enclosing scope the clone itself sits inside.
-mod clone {
-    use std::collections::BTreeMap;
-
-    use crate::ir::{Id, LinearOp, Node, Program};
-
-    use super::Builder;
-
-    pub(super) struct CloneResult {
-        /// The id reachable at explain scope, with user-iter coords appended to val.
-        pub(super) host: BTreeMap<Id, Id>,
-    }
-
-    impl Builder {
-        pub(super) fn clone_with_lifts(
-            &mut self,
-            p: &Program,
-            input_subst: &[Id],
-            import_subst: &BTreeMap<String, Id>,
-            enclosing_scope_depth: usize,
-        ) -> CloneResult {
-            let mut in_scope: BTreeMap<Id, Id> = BTreeMap::new();
-            let mut host: BTreeMap<Id, Id> = BTreeMap::new();
-            let mut user_level: usize = 0;
-            // Pending pile per enclosing user scope: (orig_id, current_cloned_id).
-            // On EndScope: lift_iter each (capturing this scope's iter coord into
-            // val), close scope, leave each to next-outer scope. If we've reached
-            // outer/explain scope, record in `host`; otherwise push onto the next-
-            // outer pile.
-            let mut pending: Vec<Vec<(Id, Id)>> = Vec::new();
-
-            for (&id, node) in &p.nodes {
-                // Exhaustive emission per node. Control-flow nodes handle
-                // their own bookkeeping in-arm and yield `None`; data nodes
-                // yield `Some(cloned_id)`, which then shares the standard
-                // in_scope/host/pending registration below.
-                let cloned: Option<Id> = match node {
-                    Node::Scope => {
-                        self.scope_open();
-                        user_level += 1;
-                        pending.push(Vec::new());
-                        None
-                    }
-                    Node::EndScope => {
-                        let pile = pending.pop().expect("EndScope without Scope");
-                        let scope_lvl = user_level; // level inside the scope
-                        // Lift each pending node BEFORE closing the scope.
-                        let lifted: Vec<(Id, Id)> = pile
-                            .into_iter()
-                            .map(|(orig, cur)| (orig, self.linear(cur, vec![LinearOp::LiftIter])))
-                            .collect();
-                        self.scope_close();
-                        user_level -= 1;
-                        // Leave each to the outer scope. `leave_dynamic(k)` truncates
-                        // the PointStamp to length `k - 1`, so we must pass the
-                        // *real* depth at this point, which is local `scope_lvl`
-                        // plus the outer offset from the enclosing explain scope.
-                        for (orig, lifted_id) in lifted {
-                            let leaved = self.leave(lifted_id, scope_lvl + enclosing_scope_depth);
-                            if user_level == 0 {
-                                host.insert(orig, leaved);
-                            } else {
-                                pending.last_mut().unwrap().push((orig, leaved));
-                            }
-                        }
-                        None
-                    }
-                    Node::Bind { variable, value } => {
-                        self.bind(in_scope[variable], in_scope[value]);
-                        None
-                    }
-                    Node::Input(i) => {
-                        // Inputs are at depth 0, host-visible directly.
-                        in_scope.insert(id, input_subst[*i]);
-                        host.insert(id, input_subst[*i]);
-                        None
-                    }
-                    Node::Import { name } => {
-                        // Imports are at depth 0, host-visible directly.
-                        let sub = *import_subst.get(name)
-                            .unwrap_or_else(|| panic!("clone: no substitution for import {:?}", name));
-                        in_scope.insert(id, sub);
-                        host.insert(id, sub);
-                        None
-                    }
-                    Node::Linear { input, ops } => {
-                        Some(self.linear(in_scope[input], ops.clone()))
-                    }
-                    Node::Concat(ids) => {
-                        let mapped: Vec<Id> = ids.iter().map(|i| in_scope[i]).collect();
-                        Some(self.concat(mapped))
-                    }
-                    Node::Arrange(input) => Some(self.arrange(in_scope[input])),
-                    Node::Join { left, right, projection } => Some(self.push(Node::Join {
-                        left: in_scope[left],
-                        right: in_scope[right],
-                        projection: projection.clone(),
-                    })),
-                    Node::Reduce { input, reducer } => Some(self.push(Node::Reduce {
-                        input: in_scope[input],
-                        reducer: reducer.clone(),
-                    })),
-                    Node::Variable => Some(self.variable()),
-                    Node::Inspect { input, label } => {
-                        Some(self.inspect(in_scope[input], label.clone()))
-                    }
-                    Node::Leave(inner, scope_level) => {
-                        Some(self.leave(in_scope[inner], *scope_level + enclosing_scope_depth))
-                    }
-                };
-
-                if let Some(cloned) = cloned {
-                    in_scope.insert(id, cloned);
-                    if user_level == 0 {
-                        // For Arrange nodes at outer scope: alias host[N] to
-                        // the underlying Collection (= host[input]) so backward
-                        // rules never refer to an Arrangement across scope
-                        // boundaries.
-                        let recorded = match node {
-                            Node::Arrange(input) => host.get(input).copied().unwrap_or(cloned),
-                            _ => cloned,
-                        };
-                        host.insert(id, recorded);
-                    } else {
-                        pending.last_mut().unwrap().push((id, cloned));
-                    }
-                }
-            }
-            assert!(pending.is_empty(), "Scope/EndScope imbalance in clone");
-
-            // Second pass: rewrite host[Leave_id] to host[inner_id] for every
-            // Leave in the program. host[inner] is the lifted form backward
-            // rules expect; the inline Leave clone above is the un-lifted form
-            // (right for `in_scope` references, wrong for `host`). Has to be a
-            // second pass because Leaves *inside* nested scopes don't have
-            // their inner's host entry populated until all enclosing scopes
-            // have closed.
-            for (&id, node) in &p.nodes {
-                if let Node::Leave(inner, _) = node {
-                    if let Some(&host_inner) = host.get(inner) {
-                        host.insert(id, host_inner);
-                    }
-                }
-            }
-
-            CloneResult { host }
-        }
-    }
-}
-
-/// Backward (demand-propagation) rule emission.
-///
-/// Per-op reverse rules all share the same skeleton: join `demand_<OUT>`
-/// with `witness + forward` of the inputs (possibly projected) on the op's
-/// natural key, filter `user_in ≤ user_out` element-wise for soundness,
-/// project to the demanded-input shape. The four `emit_lookup_*` helpers
-/// differ only in how they construct the pair table and how they map fields
-/// after the join; `filter_time_and_strip` is shared.
-mod reverse {
-    use std::collections::BTreeMap;
-
-    use crate::ir::{Id, LinearOp, Node};
-    use crate::parse::{Condition, FieldExpr, Projection, Reducer};
-
-    use super::Builder;
-    use super::CloneResult;
-
-    /// One upstream edge into a backward rule: its host-side `(data, user)`
-    /// collections from both clones, its data shape, and the two user-chain
-    /// lengths described in the top-of-module split (host vs dep). For
-    /// non-Leave nodes the two are equal; they only diverge at a `Leave`,
-    /// where the host length is the inner depth and the dep length is the
-    /// outer depth.
-    pub(super) struct Side {
-        pub(super) witness: Id,            // witness.host[input]
-        pub(super) forward: Id,            // forward.host[input]
-        pub(super) shape: (usize, usize),  // (k_arity, v_arity)
-        pub(super) host_user_len: usize,
-        pub(super) dep_user_len: usize,
-    }
-
-    impl Side {
-        pub(super) fn for_input(
-            id: Id,
-            witness: &CloneResult,
-            forward: &CloneResult,
-            arities: &BTreeMap<Id, (usize, usize)>,
-            host_user_lens: &BTreeMap<Id, usize>,
-            dep_user_lens: &BTreeMap<Id, usize>,
-        ) -> Self {
-            Self {
-                witness: witness.host[&id],
-                forward: forward.host[&id],
-                shape: arities[&id],
-                host_user_len: host_user_lens[&id],
-                dep_user_len: dep_user_lens[&id],
-            }
-        }
-    }
-
-    impl Builder {
-        pub(super) fn emit_reverse(
-            &mut self,
-            id: Id,
-            node: &Node,
-            witness: &CloneResult,
-            forward: &CloneResult,
-            demand: &BTreeMap<Id, Id>,
-            arities: &BTreeMap<Id, (usize, usize)>,
-            host_user_lens: &BTreeMap<Id, usize>,
-            dep_user_lens: &BTreeMap<Id, usize>,
-            contribs: &mut BTreeMap<Id, Vec<Id>>,
-        ) {
-            // Bind has no demand entry of its own; it routes the *variable's*
-            // demand into the *value's* contribs as a pure map.
-            //
-            // The user `var`'s forward feedback advances `user_chain[0]` (the
-            // innermost user-iter coord, which is the var's own scope iter)
-            // by 1. Inverting that: dep at `var` with `user_chain[0] = K`
-            // maps to dep at body with `user_chain[0] = K - 1`. Rows with
-            // `user_chain[0] = 0` are filtered out — they represent a demand
-            // at iter 0, which has no body-side source (the body hadn't
-            // emitted anything by then).
-            //
-            // Replacing the prior `emit_lookup_shape_preserving` (which
-            // joined against host[body] and picked up *any* body emission of
-            // the same data, regardless of iter) tightens demand to the
-            // single iter that actually sourced the var's view. Sound only
-            // because the demand variables are non-monotone — see the
-            // earlier note on the bind loop.
-            if let Node::Bind { variable, value } = node {
-                if let Some(&dv) = demand.get(variable) {
-                    let (kx, vx) = arities[variable];
-                    let var_user_len = dep_user_lens[variable];
-                    // Position of user_chain[0] in the dep row's val.
-                    let chain_pos = vx;
-                    // Filter: user_chain[0] > 0.
-                    let filtered = self.filter(
-                        dv,
-                        Condition::Gt(
-                            FieldExpr::Index(1, chain_pos),
-                            FieldExpr::Const(0),
-                        ),
-                    );
-                    // Project: subtract 1 from user_chain[0]; leave all
-                    // other fields (key, V_data, user_chain[1..], q) intact.
-                    let key: Vec<FieldExpr> =
-                        (0..kx).map(|i| FieldExpr::Index(0, i)).collect();
-                    let mut val: Vec<FieldExpr> = Vec::new();
-                    for i in 0..vx { val.push(FieldExpr::Index(1, i)); }
-                    val.push(FieldExpr::Sub(
-                        Box::new(FieldExpr::Index(1, chain_pos)),
-                        Box::new(FieldExpr::Const(1)),
-                    ));
-                    for i in 1..var_user_len {
-                        val.push(FieldExpr::Index(1, chain_pos + i));
-                    }
-                    val.push(FieldExpr::Index(1, chain_pos + var_user_len));
-                    let contrib = self.project(filtered, Projection { key, val });
-                    contribs.entry(*value).or_default().push(contrib);
-                }
-                return;
-            }
-            let dep_this = match demand.get(&id) {
-                Some(&v) => v,
-                None => return,
-            };
-            // `out_user_len` = number of user-iter coords in dep_<id>'s val (after V).
-            let out_user_len = dep_user_lens[&id];
-            let out_shape = arities[&id];
-            let side = |inp: Id| Side::for_input(inp, witness, forward, arities, host_user_lens, dep_user_lens);
-
-            match node {
-                Node::Input(_) | Node::Import { .. } => { /* terminal; feeds demand-set seeding. */ }
-
-                Node::Linear { input, ops } => {
-                    let op = match ops.as_slice() {
-                        [single] => single,
-                        _ => panic!("explain: multi-op Linear chain at node {}", id),
-                    };
-                    match op {
-                        LinearOp::Project(proj) => {
-                            let input_side = side(*input);
-                            let contrib = self.emit_lookup_lossy(dep_this, &input_side, out_shape, out_user_len, proj);
-                            contribs.entry(*input).or_default().push(contrib);
-                        }
-                        LinearOp::Filter(cond) => {
-                            // Re-apply the filter to dep_y. dep_y can carry
-                            // demand for rows the filter would have rejected
-                            // (e.g., when this Filter feeds a Concat whose
-                            // other input could have supplied those rows);
-                            // keeping only rows the filter would admit blocks
-                            // that leak. Pass-through is sound here because
-                            // input and output share a scope (Filter doesn't
-                            // cross a depth boundary; only `Leave` does).
-                            let contrib = self.filter(dep_this, cond.clone());
-                            contribs.entry(*input).or_default().push(contrib);
-                        }
-                        LinearOp::Negate => {
-                            // Pure pass-through: data unchanged, scope unchanged.
-                            contribs.entry(*input).or_default().push(dep_this);
-                        }
-                        LinearOp::EnterAt(_) => {
-                            // Sound but over-broad. EnterAt is a data→time lift: it sets a new
-                            // innermost user_chain coord `t_in = delay($field)` from the value,
-                            // so its output is one scope deeper than its input. The 1:1 reverse
-                            // would DROP that coord (recoverable as `delay($field)` from the
-                            // preserved value): out ((k,v),[t_in,t_out..]) -> in ((k,v),[t_out..]).
-                            // We instead pass demand through unchanged — tenable only because
-                            // `depths()` is positional and treats EnterAt as depth-neutral, so
-                            // the coord is stripped *unconstrained* by the neighboring Project
-                            // that crosses the scope boundary. The result is a superset (kept
-                            // sound by the `semijoin(actual_input)` at seeding); it never drops a
-                            // needed edge. Tight fix: let `depths()` give EnterAt its own level
-                            // and make this arm drop the innermost coord — see `depths()` (ir.rs).
-                            contribs.entry(*input).or_default().push(dep_this);
-                        }
-                        LinearOp::LiftIter => {
-                            // LiftIter is rewrite-emitted only; it should never
-                            // appear in a user program that reaches `explain`.
-                            panic!("explain: LiftIter in user program at node {}", id);
-                        }
-                    }
-                }
-
-                // Concat: per-input SP lookup. Each input has its own host
-                // table; the filter against `host[input]` discriminates which
-                // input "sourced" a given dep row at a given iter. Demand for
-                // (K=4, u=0) might originate from input A (host[A] has it at
-                // u=0) but not from input B (host[B] has it at u=1). Pass-
-                // through would feed demand to both, admitting B's path
-                // spuriously.
-                Node::Concat(ids) => {
-                    for inp in ids {
-                        let input_side = side(*inp);
-                        let contrib = self.emit_lookup_shape_preserving(dep_this, &input_side, out_user_len);
-                        contribs.entry(*inp).or_default().push(contrib);
-                    }
-                }
-
-                // Pure pass-through: data unchanged, scope unchanged.
-                // (Both input and output live at the same depth, so they
-                // carry the same `dep_user_len` — no depth boundary to
-                // worry about. That's `Leave`'s job, below.)
-                Node::Arrange(input) | Node::Inspect { input, .. } => {
-                    contribs.entry(*input).or_default().push(dep_this);
-                }
-
-                // Leave is the sole depth-boundary op. dep[Leave] lives
-                // in the outer scope and carries `depths[Leave]` user_chain
-                // coords, but contribs[inner] needs `depths[inner]` (one
-                // more). The SP lookup against host[inner] (which carries
-                // user_chain via `lift_iter`) injects the extra coord —
-                // and only Leave needs to know about this; no other op
-                // has to special-case "input is a Leave."
-                Node::Leave(inner, _) => {
-                    let inner_side = side(*inner);
-                    let contrib = self.emit_lookup_shape_preserving(dep_this, &inner_side, out_user_len);
-                    contribs.entry(*inner).or_default().push(contrib);
-                }
-
-                Node::Reduce { input, reducer } => {
-                    let input_side = side(*input);
-                    let contrib = self.emit_lookup_keyed(dep_this, &input_side, out_shape, out_user_len, reducer);
-                    contribs.entry(*input).or_default().push(contrib);
-                }
-
-                Node::Join { left, right, projection } => {
-                    let left_side = side(*left);
-                    let right_side = side(*right);
-                    let (left_contrib, right_contrib) = self.emit_lookup_join(
-                        dep_this, &left_side, &right_side,
-                        out_shape, out_user_len, projection,
-                    );
-                    contribs.entry(*left).or_default().push(left_contrib);
-                    contribs.entry(*right).or_default().push(right_contrib);
-                }
-
-                Node::Variable => { /* handled by Bind. */ }
-
-                Node::Bind { .. } | Node::Scope | Node::EndScope => {}
-            }
-        }
-
-    /// Shape-preserving lookup: input and output have the same (k, v) shape.
-    /// Used by Concat — each input gets its own host lookup so the per-input
-    /// time filter can discriminate which input sourced a given dep row.
-    ///
-    /// pair = witness + forward of input. Shape (K; V ++ user_chain_in).
-    /// Repack to (K+V; user_chain_in). dep repacks to (K+V; user_chain_out ++ [q]).
-    /// Join on (K+V); produce (K; V ++ user_chain_in ++ user_chain_out ++ [q]),
-    /// then `filter_time_and_strip` enforces `user_in ≤ user_out` and drops
-    /// `user_chain_out` to yield `(K; V ++ user_chain_in ++ [q])`.
-    fn emit_lookup_shape_preserving(
-        &mut self,
-        dep_y: Id,
-        side: &Side,
-        output_depth: usize,
-    ) -> Id {
-        let (k, v) = side.shape;
-        // `host_user_len` is the user_chain length in host[input] — what the
-        // pair table actually carries. `dep_user_len` is what contribs[input]
-        // should carry (= depths[input]). They differ only when `input` is a
-        // `Leave`; the lookup truncates user_in down to `dep_user_len` on the
-        // way out.
-        let host_user_len = side.host_user_len;
-        let dep_user_len = side.dep_user_len;
-        let pair = self.concat(vec![side.witness, side.forward]);
-        let pair_keyed = self.project(pair, Projection {
-            key: pack_kv(k, v),
-            val: (0..host_user_len).map(|i| FieldExpr::Index(1, v + i)).collect(),
-        });
-        let dep_keyed = self.project(dep_y, Projection {
-            key: pack_kv(k, v),
-            val: (0..output_depth + 1).map(|i| FieldExpr::Index(1, v + i)).collect(),
-        });
-        // After arrange-join on (K+V): $0 = K+V, $1 = dep val, $2 = pair val.
-        // Keep user_out in val for the time filter, then strip.
-        let key: Vec<FieldExpr> = (0..k).map(|i| FieldExpr::Index(0, i)).collect();
-        let mut val: Vec<FieldExpr> = Vec::new();
-        for i in 0..v { val.push(FieldExpr::Index(0, k + i)); }
-        for i in 0..host_user_len { val.push(FieldExpr::Index(2, i)); }
-        for i in 0..output_depth { val.push(FieldExpr::Index(1, i)); }
-        val.push(FieldExpr::Index(1, output_depth)); // q
-        let joined = self.join(dep_keyed, pair_keyed, Projection { key, val });
-        self.filter_time_and_strip(joined, k, v, host_user_len, output_depth, dep_user_len)
-    }
-
-    /// Apply the soundness filter (`user_in[i] ≤ user_out[i]` element-wise) and
-    /// strip `user_out` from a collection whose row shape is
-    /// `(K[k_out]; V_pre[v_pre] ++ user_in[in_len] ++ user_out[out_len] ++ [q])`.
-    /// Result shape: `(K[k_out]; V_pre[v_pre] ++ user_in[0..keep_in_len] ++ [q])`.
-    ///
-    /// The filter excludes contributions whose witness input row was produced at
-    /// a strictly-later user-iter than the demanded output — an output cannot be
-    /// "explained by" an input that came after it. When `in_len` and `out_len`
-    /// differ, we compare only the common prefix (innermost-first ordering).
-    ///
-    /// `keep_in_len` lets callers ask for fewer user_in coords in the output
-    /// than the pair table carries. This is how a `Leave`-as-input collapses
-    /// host[Leave]'s inner user_chain (length `depths[inner]`) down to
-    /// contribs[Leave]'s outer user_chain (length `depths[Leave]`).
-    fn filter_time_and_strip(
-        &mut self,
-        coll: Id,
-        k_out: usize,
-        v_pre: usize,
-        in_len: usize,
-        out_len: usize,
-        keep_in_len: usize,
-    ) -> Id {
-        // The user_chain index arithmetic (the outer-end alignment whose
-        // off-by-one made SCC explain unsound) lives in `folded::Joined`, the
-        // single home shared with the scope-builder ports.
-        let layout = crate::folded::Joined { v_pre, in_len, out_len };
-        let mut cur = coll;
-        if let Some(cond) = layout.time_le() {
-            cur = self.filter(cur, cond);
-        }
-        self.project(cur, layout.strip(k_out, keep_in_len))
-    }
-
-    /// Keyed lookup (Reduce-style): demand on `(K; V_out ++ user_out ++ q)`
-    /// maps to every input row at the same K, time-filtered against user_out.
-    /// Output: `(K; V_in ++ user_in ++ [q])`.
-    ///
-    /// For `Reducer::Min`, also applies the §3.2.2 value-narrowing: only
-    /// keep input rows whose `V_in` equals the queried `V_out` (the min).
-    /// Other inputs at the same key did not contribute to that min and are
-    /// not needed in the explanation.
-    fn emit_lookup_keyed(
-        &mut self,
-        dep_y: Id,
-        side: &Side,
-        output_shape: (usize, usize),
-        out_user_len: usize,
-        reducer: &Reducer,
-    ) -> Id {
-        let (k_in, v_in) = side.shape;
-        let in_user_len = side.host_user_len;
-        let keep_in_len = side.dep_user_len;
-        let (_, v_out) = output_shape;
-        let pair = self.concat(vec![side.witness, side.forward]);
-        // Min narrowing: include V_out in the val layout so we can filter
-        // V_in == V_out element-wise below. Only applies when arities match
-        // (true for Min: forward preserves arity) and there's a value to compare.
-        let include_v_out = matches!(reducer, Reducer::Min) && v_in == v_out && v_in > 0;
-        // After arrange-join on K: $0 = K, $1 = dep val (V_out + user_out + q),
-        // $2 = pair val (V_in + user_in).
-        // val layout: V_in [+ V_out for Min] + user_in + user_out + q.
-        let mut val: Vec<FieldExpr> = Vec::new();
-        for i in 0..v_in { val.push(FieldExpr::Index(2, i)); }                      // V_in
-        if include_v_out {
-            for i in 0..v_out { val.push(FieldExpr::Index(1, i)); }                 // V_out
-        }
-        for i in 0..in_user_len { val.push(FieldExpr::Index(2, v_in + i)); }        // user_in
-        for i in 0..out_user_len { val.push(FieldExpr::Index(1, v_out + i)); }      // user_out
-        val.push(FieldExpr::Index(1, v_out + out_user_len));                         // q
-        let proj = Projection {
-            key: (0..k_in).map(|i| FieldExpr::Index(0, i)).collect(),
-            val,
-        };
-        let joined = self.join(dep_y, pair, proj);
-
-        // For Min: filter V_in[i] == V_out[i] element-wise, then strip V_out
-        // back out of val so the downstream time-filter sees its expected
-        // (K; V_in + user_in + user_out + q) layout.
-        let after_min = if include_v_out {
-            let mut acc: Option<Condition> = None;
-            for i in 0..v_in {
-                let cond = Condition::Eq(
-                    FieldExpr::Index(1, i),         // V_in[i]
-                    FieldExpr::Index(1, v_in + i),  // V_out[i]
-                );
-                acc = Some(match acc {
-                    None => cond,
-                    Some(prev) => Condition::And(Box::new(prev), Box::new(cond)),
-                });
-            }
-            let filtered = self.filter(joined, acc.unwrap());
-            // Re-project to drop V_out, restoring (K; V_in + user_in + user_out + q).
-            let key: Vec<FieldExpr> = (0..k_in).map(|i| FieldExpr::Index(0, i)).collect();
-            let mut new_val: Vec<FieldExpr> = Vec::new();
-            for i in 0..v_in { new_val.push(FieldExpr::Index(1, i)); }                       // V_in
-            // skip V_out at [v_in..v_in+v_out]
-            let after_vout = v_in + v_out;
-            for i in 0..in_user_len { new_val.push(FieldExpr::Index(1, after_vout + i)); }   // user_in
-            for i in 0..out_user_len { new_val.push(FieldExpr::Index(1, after_vout + in_user_len + i)); }  // user_out
-            new_val.push(FieldExpr::Index(1, after_vout + in_user_len + out_user_len));      // q
-            self.project(filtered, Projection { key, val: new_val })
-        } else {
-            joined
-        };
-
-        self.filter_time_and_strip(after_min, k_in, v_in, in_user_len, out_user_len, keep_in_len)
-    }
-
-    /// Lossy lookup (Linear[Project]).
-    ///
-    /// Two emission paths depending on `proj`'s invertibility:
-    ///
-    /// **Pure-map shortcut.** If every input field is recoverable from the
-    /// output (every `(r, c)` appears as `Index(r, c)` or via `Pos(r)` in
-    /// `proj.key` or `proj.val`) AND the input lives at outer scope (no
-    /// `user_chain` to source), bypass the pair table entirely: emit a
-    /// single Linear that maps each `dep_y` row directly to its synthetic
-    /// `(K_in, V_in, q)` form. Cost is `O(|dep|)`, independent of `|input|`
-    /// — the algorithmic win that lets explanations stay bounded as inputs
-    /// grow. Any synthetic input row that doesn't actually exist gets
-    /// filtered out later by the `semijoin(actual_input_<i>)` at must-set
-    /// seeding.
-    ///
-    /// **Fallback (pair-table).** Otherwise, the original construction:
-    /// `pair = (witness_input + forward_input)` projected to
-    /// `(chained_K_out; K_in ++ V_in ++ user_in)`. dep_y has shape
-    /// `(K_out; V_out ++ user_out ++ q)`. After join on K_out, output
-    /// `(K_in; V_in ++ user_in ++ [q])`, time-filtered.
-    fn emit_lookup_lossy(
-        &mut self,
-        dep_y: Id,
-        side: &Side,
-        output_shape: (usize, usize),
-        out_user_len: usize,
-        proj: &Projection,
-    ) -> Id {
-        let (k_in, v_in) = side.shape;
-        let in_user_len = side.host_user_len;
-        let keep_in_len = side.dep_user_len;
-        let (k_out, v_out) = output_shape;
-
-        let known = analyze_lossy_invertibility(proj, k_in, v_in);
-        let total = (0..k_in).all(|c| known.contains_key(&(0, c)))
-            && (0..v_in).all(|c| known.contains_key(&(1, c)));
-
-        // Pure-map shortcut, extended to same-scope projects of any depth.
-        // A Linear[Project] doesn't cross a scope boundary, so the input's
-        // user_chain equals the output's (in_user_len == out_user_len ==
-        // keep_in_len when the input isn't a Leave). When every input field is
-        // also recoverable from the output, the whole reverse is a direct map
-        // from dep_y, narrowing by *all* demanded fields — including value
-        // fields the pair-table join (keyed on K_out only) ignores. That
-        // key-only match is what let sibling rows leak through a re-key that
-        // drops a field from the key (e.g. by_b = tc | key($0[1]; …), keyed by
-        // tc's destination only → fallback recovered every same-destination
-        // pair regardless of source).
-        if total && in_user_len == out_user_len && keep_in_len == in_user_len {
-            // Map flat output position p (into [K_out ++ V_out]) to an access
-            // expression against dep_y's (key, val) layout:
-            //   p < k_out → $0[p]                 (key)
-            //   p >= k_out → $1[p - k_out]        (val, since dep_y.val starts with V_out)
-            let access = |p: usize| -> FieldExpr {
-                if p < k_out { FieldExpr::Index(0, p) }
-                else { FieldExpr::Index(1, p - k_out) }
-            };
-            let key: Vec<FieldExpr> = (0..k_in).map(|c| access(known[&(0, c)])).collect();
-            let mut val: Vec<FieldExpr> = Vec::with_capacity(v_in + in_user_len + 1);
-            for c in 0..v_in { val.push(access(known[&(1, c)])); }
-            // user_in == user_out (same scope), so copy it through. The filter
-            // user_in ≤ user_out then holds with equality. q at $1[v_out+out_user_len].
-            for i in 0..out_user_len { val.push(FieldExpr::Index(1, v_out + i)); }
-            val.push(FieldExpr::Index(1, v_out + out_user_len));
-            return self.project(dep_y, Projection { key, val });
-        }
-
-        let pair_src = self.concat(vec![side.witness, side.forward]);
-        // pair_src shape: (K_in; V_in ++ user_in[0..in_user_len]).
-        // Build pair: key = proj.key (computes chained K_out), val = K_in ++ V_in ++ user_in.
-        let mut pair_val: Vec<FieldExpr> = Vec::with_capacity(k_in + v_in + in_user_len);
-        for i in 0..k_in { pair_val.push(FieldExpr::Index(0, i)); }
-        for i in 0..v_in + in_user_len { pair_val.push(FieldExpr::Index(1, i)); }
-        let pair = self.project(pair_src, Projection { key: proj.key.clone(), val: pair_val });
-        // After arrange-join on K_out: $0 = K_out, $1 = dep val (V_out + user_out + q),
-        // $2 = pair val (K_in + V_in + user_in).
-        // Keep user_out in val for the time filter, then strip.
-        let key: Vec<FieldExpr> = (0..k_in).map(|i| FieldExpr::Index(2, i)).collect();
-        let mut val: Vec<FieldExpr> = Vec::new();
-        for i in 0..v_in { val.push(FieldExpr::Index(2, k_in + i)); }
-        for i in 0..in_user_len { val.push(FieldExpr::Index(2, k_in + v_in + i)); }
-        for i in 0..out_user_len { val.push(FieldExpr::Index(1, v_out + i)); }
-        val.push(FieldExpr::Index(1, v_out + out_user_len));
-        let joined = self.join(dep_y, pair, Projection { key, val });
-        self.filter_time_and_strip(joined, k_in, v_in, in_user_len, out_user_len, keep_in_len)
-    }
-
-    /// Join's backward rule: produces two contribs (left and right).
-    fn emit_lookup_join(
-        &mut self,
-        dep_y: Id,
-        left: &Side,
-        right: &Side,
-        output_shape: (usize, usize),
-        out_user_len: usize,
-        projection: &Projection,
-    ) -> (Id, Id) {
-        let (k_arity, v_l) = left.shape;
-        let (_, v_r) = right.shape;
-        let (_, v_out) = output_shape;
-        let left_user_len = left.host_user_len;
-        let right_user_len = right.host_user_len;
-        let left_keep_in_len = left.dep_user_len;
-        let right_keep_in_len = right.dep_user_len;
-        // Left and right inputs have shape (K; V_L/R ++ user_L/R[user_len]).
-        let left_pair_src = self.concat(vec![left.witness, left.forward]);
-        let right_pair_src = self.concat(vec![right.witness, right.forward]);
-        // Forward-join them on K, projecting to
-        //   (chained_K_out; K ++ V_L ++ V_R ++ user_L ++ user_R).
-        // In the join's projection $0 = K, $1 = left val (V_L + user_L), $2 = right val.
-        let mut pair_val: Vec<FieldExpr> = Vec::new();
-        for i in 0..k_arity { pair_val.push(FieldExpr::Index(0, i)); }
-        for i in 0..v_l { pair_val.push(FieldExpr::Index(1, i)); }
-        for i in 0..v_r { pair_val.push(FieldExpr::Index(2, i)); }
-        for i in 0..left_user_len { pair_val.push(FieldExpr::Index(1, v_l + i)); }
-        for i in 0..right_user_len { pair_val.push(FieldExpr::Index(2, v_r + i)); }
-        // The user's projection.key may reference `$1` / `$2` as whole-row
-        // expansions (FieldExpr::Pos), which is correct against the *original*
-        // V_L / V_R but wrong against the lift-extended host V's (which append
-        // user-iter coords). Rewrite Pos(i) -> bounded Index(i, 0..arity_i) so
-        // the projection only sees the original-shape fields and yields the
-        // same K_out as the user's forward join.
-        let pos_arities = [k_arity, v_l, v_r];
-        let key_expanded = expand_pos_bounded(&projection.key, &pos_arities);
-        let pair = self.join(
-            left_pair_src,
-            right_pair_src,
-            Projection { key: key_expanded, val: pair_val },
-        );
-        // pair val arity: k_arity + v_l + v_r + left_user_len + right_user_len.
-        // After arrange-join with dep_y on K_out:
-        //   $0 = K_out, $1 = dep val (v_out + out_user_len + 1),
-        //   $2 = pair val (k_arity + v_l + v_r + left_user_len + right_user_len).
-        let q_pair_pos = v_out + out_user_len;
-        let k_pair_start = 0;
-        let vl_pair_start = k_pair_start + k_arity;
-        let vr_pair_start = vl_pair_start + v_l;
-        let ul_pair_start = vr_pair_start + v_r;
-        let ur_pair_start = ul_pair_start + left_user_len;
-        // Left contrib: (K; V_L + user_L + user_out + [q]), then filter+strip.
-        let key_left: Vec<FieldExpr> =
-            (0..k_arity).map(|i| FieldExpr::Index(2, k_pair_start + i)).collect();
-        let mut val_left: Vec<FieldExpr> = Vec::new();
-        for i in 0..v_l { val_left.push(FieldExpr::Index(2, vl_pair_start + i)); }
-        for i in 0..left_user_len { val_left.push(FieldExpr::Index(2, ul_pair_start + i)); }
-        for i in 0..out_user_len { val_left.push(FieldExpr::Index(1, v_out + i)); }
-        val_left.push(FieldExpr::Index(1, q_pair_pos));
-        let left_joined = self.join(dep_y, pair, Projection { key: key_left, val: val_left });
-        let left_contrib = self.filter_time_and_strip(left_joined, k_arity, v_l, left_user_len, out_user_len, left_keep_in_len);
-        // Right contrib: (K; V_R + user_R + user_out + [q]), then filter+strip.
-        let key_right: Vec<FieldExpr> =
-            (0..k_arity).map(|i| FieldExpr::Index(2, k_pair_start + i)).collect();
-        let mut val_right: Vec<FieldExpr> = Vec::new();
-        for i in 0..v_r { val_right.push(FieldExpr::Index(2, vr_pair_start + i)); }
-        for i in 0..right_user_len { val_right.push(FieldExpr::Index(2, ur_pair_start + i)); }
-        for i in 0..out_user_len { val_right.push(FieldExpr::Index(1, v_out + i)); }
-        val_right.push(FieldExpr::Index(1, q_pair_pos));
-        let right_joined = self.join(dep_y, pair, Projection { key: key_right, val: val_right });
-        let right_contrib = self.filter_time_and_strip(right_joined, k_arity, v_r, right_user_len, out_user_len, right_keep_in_len);
-        (left_contrib, right_contrib)
-    }
-    }
-
-    /// Rewrite `FieldExpr::Pos(i)` in a key/val list to a bounded expansion
-    /// `[Index(i, 0), Index(i, 1), ..., Index(i, arities[i]-1)]`. Used in
-    /// `emit_lookup_join` where the user's projection is applied to host-side
-    /// (lift-extended) inputs — `Pos(i)` against the extended row would also
-    /// pick up the trailing user-iter coords, which we explicitly do not want
-    /// inside the key.
-    fn expand_pos_bounded(fields: &[FieldExpr], arities: &[usize]) -> Vec<FieldExpr> {
-        let mut out = Vec::with_capacity(fields.len());
-        for f in fields { expand_pos_one(f, arities, &mut out); }
-        out
-    }
-
-    fn expand_pos_one(f: &FieldExpr, arities: &[usize], out: &mut Vec<FieldExpr>) {
-        match f {
-            FieldExpr::Pos(i) => {
-                for c in 0..arities[*i] { out.push(FieldExpr::Index(*i, c)); }
-            }
-            FieldExpr::Index(_, _) | FieldExpr::Const(_) => out.push(f.clone()),
-            FieldExpr::Neg(inner) => {
-                let mut tmp = Vec::new();
-                expand_pos_one(inner, arities, &mut tmp);
-                for t in tmp { out.push(FieldExpr::Neg(Box::new(t))); }
-            }
-            FieldExpr::Sub(a, b) => {
-                let mut ta = Vec::new();
-                let mut tb = Vec::new();
-                expand_pos_one(a, arities, &mut ta);
-                expand_pos_one(b, arities, &mut tb);
-                for (x, y) in ta.into_iter().zip(tb.into_iter()) {
-                    out.push(FieldExpr::Sub(Box::new(x), Box::new(y)));
-                }
-            }
-        }
-    }
-
-    fn pack_kv(k: usize, v: usize) -> Vec<FieldExpr> {
-        let mut out: Vec<FieldExpr> = Vec::with_capacity(k + v);
-        for i in 0..k { out.push(FieldExpr::Index(0, i)); }
-        for i in 0..v { out.push(FieldExpr::Index(1, i)); }
-        out
-    }
-
-    /// Identify which input fields a `Linear[Project]` projection makes
-    /// recoverable from its output. Returns a map `(r, c) → output_position`
-    /// where `output_position` indexes into the flattened `[K_out, V_out]`
-    /// row and `(r, c)` is the input field — `r=0` for `K_in[c]`, `r=1` for
-    /// `V_in[c]`.
-    ///
-    /// Direct cases (`Index(r, c)` and `Pos(r)`) are recognized. Computed
-    /// fields (`Const`, `Neg`, `Sub`) are not analyzed — the input field
-    /// they reference, if any, is left "unknown" and falls back to the
-    /// pair-table path in `emit_lookup_lossy`. This is sound but loose.
-    ///
-    /// Tightening opportunities for whoever adds a parser entry for these:
-    /// - `Neg(Index(r, c))` recovers input via `-output_at_p`.
-    /// - `Sub(Index(r, c), Const(k))` recovers via `output_at_p + k`.
-    /// - `Sub(Const(k), Index(r, c))` recovers via `k - output_at_p`.
-    /// Recording these as recoverable would let the pure-map shortcut fire
-    /// for projections that shift fields by a constant. Today none of the
-    /// rewrite's own emissions or user programs reach this site with such
-    /// expressions, so it's purely latent looseness.
-    fn analyze_lossy_invertibility(
-        proj: &Projection,
-        k_in: usize,
-        v_in: usize,
-    ) -> BTreeMap<(usize, usize), usize> {
-        let mut known: BTreeMap<(usize, usize), usize> = BTreeMap::new();
-        let mut p: usize = 0;
-        for fe in proj.key.iter().chain(proj.val.iter()) {
-            match fe {
-                FieldExpr::Index(r, c) => {
-                    known.entry((*r, *c)).or_insert(p);
-                    p += 1;
-                }
-                FieldExpr::Pos(r) => {
-                    let arity = if *r == 0 { k_in } else { v_in };
-                    for c in 0..arity {
-                        known.entry((*r, c)).or_insert(p + c);
-                    }
-                    p += arity;
-                }
-                FieldExpr::Const(_) | FieldExpr::Neg(_) | FieldExpr::Sub(_, _) => {
-                    p += 1;
-                }
-            }
-        }
-        known
-    }
-}
-
-// Projection helpers (used by `explain` itself, not by reverse rules).
-
-impl Builder {
-    /// Insert an `Inspect` node for diagnostic output, but only when the
-    /// `EXPLAIN_DEBUG_DEP` env var is set. Used to surface dep / pair
-    /// tables at construction sites without bloating the IR in normal runs.
-    pub fn debug_inspect(&mut self, input: Id, label: String) {
-        if std::env::var("EXPLAIN_DEBUG_DEP").is_ok() {
-            self.inspect(input, label);
-        }
-    }
-
-    /// Semijoin `left (K; V)` with `right (K; V)` by `(K)`, keep left's rows.
-    /// (Used at demand-set seeding.)
-    pub fn semijoin_data(&mut self, left: Id, right: Id, k_arity: usize, v_arity: usize) -> Id {
-        let proj_key: Vec<FieldExpr> = (0..k_arity).map(|i| FieldExpr::Index(0, i)).collect();
-        let proj_val: Vec<FieldExpr> = (0..v_arity).map(|i| FieldExpr::Index(1, i)).collect();
-        self.join(left, right, Projection { key: proj_key, val: proj_val })
-    }
-
-    /// Set-level distinct on `(K; V)` rows (DDIR `distinct` is per-key only;
-    /// pack-distinct-unpack preserves the val).
-    pub fn distinct_full(&mut self, input: Id, k_arity: usize, v_arity: usize) -> Id {
-        let mut pack_key: Vec<FieldExpr> = (0..k_arity).map(|i| FieldExpr::Index(0, i)).collect();
-        for i in 0..v_arity {
-            pack_key.push(FieldExpr::Index(1, i));
-        }
-        let packed = self.project(input, Projection { key: pack_key, val: vec![] });
-        let dist = self.reduce(packed, Reducer::Distinct);
-        let unpack_key: Vec<FieldExpr> = (0..k_arity).map(|i| FieldExpr::Index(0, i)).collect();
-        let unpack_val: Vec<FieldExpr> = (0..v_arity).map(|i| FieldExpr::Index(0, k_arity + i)).collect();
-        self.project(dist, Projection { key: unpack_key, val: unpack_val })
-    }
-}
-
-/// Strip user-chain and `[q]` from a dep row's val: `(K; V ++ user ++ [q])` -> `(K; V)`.
-fn strip_user_and_q(k_arity: usize, v_arity: usize) -> Projection {
-    let key: Vec<FieldExpr> = (0..k_arity).map(|i| FieldExpr::Index(0, i)).collect();
-    let val: Vec<FieldExpr> = (0..v_arity).map(|i| FieldExpr::Index(1, i)).collect();
-    Projection { key, val }
-}
-
