+++ interactive/src/lib.rs
@@ -1,7 +1,6 @@
 pub mod parse;
 pub mod ir;
 pub mod lower;
-pub mod explain;
 pub mod folded;
 pub mod scope_ir;
 pub mod explain_tree;
